Replace my older fork of `ahoy` with new release
I'm a maintainer of `ahoy` now, and we have released an update, version 2.0.1 (more accurate semver than my fork was). There are no breaking changes, only some Go module updates and a shift to a newer Go version for building. Further improvements and bugfixes will follow, and will be properly versioned with semver.